reorg and build against newest sl_sh_proc_macro crate
